fix generator expression for setuptools to find the package
forgetting this comma but providing the MODULE_NAME string in a generator expression most likely means setuptools looked for t, u, r, n, ... etc when performing the install instead of turn_by_turn and installed none of this.

With the comma, we force MODULE_NAME to be a single element, which is our package's name to be installed. This lets users properly install and import turn_by_turn.

unfortunately warrants a new release